Extract about content component
