+++ src/main/java/io/fxgame/game2048/Board.java
@@ -19,12 +19,10 @@ import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
-import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.Group;
 import javafx.scene.control.Button;
 import javafx.scene.control.ChoiceBox;
-import javafx.scene.control.Hyperlink;
 import javafx.scene.control.Label;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Pane;
@@ -32,9 +30,6 @@ import javafx.scene.layout.Priority;
 import javafx.scene.layout.VBox;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Rectangle;
-import javafx.scene.text.Text;
-import javafx.scene.text.TextAlignment;
-import javafx.scene.text.TextFlow;
 import javafx.util.Duration;
 
 /**
@@ -411,48 +406,8 @@ public class Board extends Pane {
 
     private void showAboutOverlay() {
         timer.stop();
-        overlayPanel.showContent("game-overlay-quit", bContinue, null, createAboutContent());
+        overlayPanel.showContent("game-overlay-quit", bContinue, null, new AboutContent(gridDimension));
         showOverlay(bContinue);
-    }
-
-    private TextFlow createAboutContent() {
-        var flow = new TextFlow();
-        flow.setTextAlignment(TextAlignment.CENTER);
-        flow.setPadding(new Insets(10, 0, 0, 0));
-        flow.setMinSize(gridDimension, gridDimension);
-        flow.setPrefSize(gridDimension, gridDimension);
-        flow.setMaxSize(gridDimension, gridDimension);
-        flow.setPrefSize(Double.NEGATIVE_INFINITY, Double.NEGATIVE_INFINITY);
-
-        flow.getChildren().setAll(
-                styledText("2048", "game-lblAbout"),
-                styledText("FX", "game-lblAbout2"),
-                styledText(" Game\n", "game-lblAbout"),
-                styledText("JavaFX game - Desktop version\n\n", "game-lblAboutSub"),
-                styledText("Powered by ", "game-lblAboutSub"),
-                aboutLink("OpenJFX", "https://openjfx.io/"),
-                styledText(" Project \n\n", "game-lblAboutSub"),
-                styledText("© ", "game-lblAboutSub"),
-                aboutLink("@JPeredaDnr", "https://twitter.com/JPeredaDnr"),
-                styledText(" & ", "game-lblAboutSub"),
-                aboutLink("@brunoborges", "https://twitter.com/brunoborges"),
-                styledText("\n\n", "game-lblAboutSub"),
-                styledText("Version " + Game2048.VERSION + "\n\n", "game-lblAboutSub"));
-
-        return flow;
-    }
-
-    private Text styledText(String text, String styleClass) {
-        var styledText = new Text(text);
-        styledText.getStyleClass().setAll("game-label", styleClass);
-        return styledText;
-    }
-
-    private Hyperlink aboutLink(String text, String url) {
-        var link = new Hyperlink(text);
-        link.setOnAction(_ -> Game2048.urlOpener().open(url));
-        link.getStyleClass().setAll("game-label", "game-lblAboutSub2");
-        return link;
     }
 
     public void animateScore() {
